Restructure how we handle request parsing.
With this, we parse the raw body immediately, and use a preserving
supplier to buffer the *output* of the raw body parsing process for
a downstream body parser. This means we can rely on having `$leftover`
reflect the outcome of the parse immediately after pushing the packet
into the raw parsing pipeline.

There is one further simpliciation to the logic here. We don't need to
special-case when the blob that `$leftover` was kept with was 0 length
and can just let that fall out of the fact that there will be no line
available from the decoder, which will do the `last` for us in the
next iteration.