Basic implementation of delegate
This works for the basic case, but won't yet work for using it to do a
delegation to another route block, because we don't yet have any notion
of "effective target" and "effective path segments". We also only ever
delegate to an exact match, rather than allow extra segments; need to
consider and decide on the semantics here.