Initial router support for auth
This implements the case where a Cro::HTTP::Auth typed parameter is
used at the start of the route segment. Note that a delegate is
required here for now because the route matcher currently runs ahead
of any middleware. That may or may not be something we wish to revisit
in the future.