Refactor handling of push promises in responses
Get the response parser to attach them, which is simpler. This fixes
the hang in the client getting push promises (which was caused by the
client itself tapping the preserved Supply of them and thus eating
them, so the end user never got to see them). Along with this, turn
the test file exhibiting the hang into some basic push promise tests.