Make `:@query` param reflect query string order
Previously we made a hash of query string parameters and then just did
`.List` to make it into a list of `Pair`. This made it vulnerable to
the recent hash order changes in Perl 6. However, it's probably more
useful to have the list form preserve the exact order found in the
query string, as well as not joining multiple same-named parameters.
This fixes the test, and makes it consistent with `:@foo is header`,
which also retains order and doesn't join multiple values.