Add a uri method to Cro::HTTP request
This gives the full request URI. When it's a client-side request, we
have the issued request URI to hand and so can use that. On the server
side, we reconstruct it best we can.

Along the way, I noticed that we weren't compliant with the HTTP spec
on the host header, which indicates that if the port the request was
made on is a non-standard one for HTTP, then it should be included in
the Host header. That is also fixed and test-covered with this commit.

This satisfies the request in issue #24; the solution needs to be
documented.