Take a 400/401 error over a 405 error
This resolves #59. When route matching, we might find multiple routes
that all fail for some reason. If we find one that fails to bind check,
that means it matched segments and HTTP method, and failed *after*
that. Thus, it "got further" than a route that simply fails to match on
HTTP method, and its error should be picked in preference.