Update default TLS cipher list
Mozilla have updated their TLS recommendations. We can no longer use the "Modern" preset as I don't think we support TLS 1.3 yet, but the "Intermediate" preset is probably a better fit as it's the recommended preset for a general-purpose server. The main change is the dropping of CBC in favour of GCM. This change restores a SSL Labs rating of "A" out of the box.