Emit HTTP/2 request/response once headers are done
Don't wait for the whole body to be produced. This is how things should
work anyway: it's better for throughput, and makes using HTTP/2 in
streaming body cases possible. But it also was the reason for the hang
reported in #85: there's an error after the headers are received, which
was reported on the body supply, but nobody could ever possibly get to
that because we never emitted the response object. (There's still a
further issue that needs exploring for #85, but this means we at least
get an error out now, not a hang.)