Support "Transfer-Encoding: identity"
Use Content-Length or otherwise to determine the RawBodyParser as
the 'identity' transfer-encoding implies 'as-is' encoding.

RFC 2616 indicates https://tools.ietf.org/html/rfc2616#section-3.5 that
this SHOULD NOT be used in a Transfer-Encoding header, but there is at
least one library that does emit this (as noted in #89 ,) so it seems
prudent to accept it even if it is non-compliant.

Fixes #89