Handle capture-related exceptions in request-body gracefully
It turns out ACCEPTS could throw an exception instead of silently
returning False when it cannot unpack something.

It is an ambiguous error we should not show to the client, so
better hide it behind a 400, but in fact it can be a developer's
mis-understanding somewhere as well, so report them some info
that can be useful in debugging.