Provide get-body and friends in Cro::HTTP::Client
Very often one only wants the response body and doesn't care about the
headers. However, today we make people write out this boilerplate:

    my $response = await $client.get('http://foo.bar/baz');
    my $body = await $response.body;

Provide a `get-body` convenience method, which reduces it to:

    my $body = await $client.get-body('http://foo.bar/baz');

This is done for the other HTTP methods, except `HEAD`, since that is
explicitly requesting only the headers, so a `head-body` method would be
rather silly.