Provide working serving of resources
There was a previous attempt - `static-resource` - however, it did not
work correctly, because %?RESOURCES always looks for the resources of
the current distribution, which meant it would always look at the
Cro::HTTP distribution's resources. This meant that it passed the tests,
but that was about it.

For this to work reliably, we should explicitly associate the
%?RESOURCES with the `route` block. For this, `resources-from` is
introduced, which uses the router plugin mechaism to store the resources
hash against the `route` block (meaning the right thing will happen with
`include`). Then, the sub `resource` is used to serve resources. The
non-working `static-resource` is gone; it was a bit wordy for something
that may well be quite common going forward.

The new test case uses a distinct module to try and create a more
realistic example; as a side benefit, this means we don't install the
testing resources as part of Cro::HTTP.