Provide for other plugins wanting resources
We'll also make it so that Cro::WebApp::Template can serve resources.
However, it would be nice if we only had to associate the resources with
the routes using `resources-from` once for both serving static files and
for templates; templates will then need a single extra opt-in to say
that resources should be considered. Therefore, expose a sub that can be
used to resolve a resource, so Cro::WebApp::Template can use that.