Get response body timeout working correctly
The previous approach didn't correctly factor the time elapsed for the
connection and headers phase into the total, and also would not handle
persistent connections correctly. Do better by providing a mechanism for
a HTTP response to be cancelled - that is to say, terminated even before
the body has been fully received. Use this to enforce the timeout (and
gain a bonus feature along with it). Only implemented for HTTP/1.1 thus
far.