Implement remote window handling
The basic idea is to provide a mechanism of flow control back pressure, so
the remote end is not flodded with data it can't process fast enough.
The basic mechanism is as follows: We have a contingent of bytes we are
allowed to send in DATA frames to the remote party. If that contingent is
depleted we have to wait for the remote end to refresh it again via
WINDOW_UPDATE frames before we can continue to send more DATA frames.