Fix hash issues with TLS options changes
The changes merged in 50cf7987b led to some tests failing with the
error "Odd number of elements found where hash initializer expected".
This resolves the error and also makes sure that options passed with a
request take priority over the class-level settings, for consistency
with how options handling works elsewhere in the client.