Implement a retry mechanism when GoAways are received
Remote side connection tear down (via a GoAway packet) and new requests
are racey. If they collide the request fails. The user shouldn't be
bothered to handle that case. Thus we automatically retry the request
should a request fail as a result of a GoAway race condition.

Also GoAways are now communicated from the GeneralParser to HTTP::Client
via Exceptions instead of a separate supply this simplifies the interface.

Hint for testing this:

I used a local Apache webserver and introduced artificial latency on the
loopback device via

    tc qdisc add dev lo root netem delay 100ms

Then doing two consecutive requests on the same `Cro::HTTP::Client` object
with a tuned sleep in between manages to hit the race condition.