Fix another connection reuse race
Connection reuse is subject to a race condition where remote connection
tear down races with reusing the connection for a new request. There were
already locks in place for access to `$!next-response-vow` /
`%!outstanding-stream-responses` but the lock didn't cover the check for
`$!dead`, thus it was possible for the dead check to succeed, the
connection being torn down and only then the next response vow to be taken.