Fix HTTP2 connections with multiple streams stalling
HTTP2 knows a stream and connection flow-control window. Both windows need
to be respected. A logic was put in place, to not update the window sizes
when the stream that sent the data was closed. That's fine. But we do need
to update the global window, otherwise the global window size permanently
shrinks by that amount. When reusing a connection many times, this shinks
the window size down to zero over time, effectively stalling the
connection.