Adding named endpoints
It adds `Str :$name` to the `http` functions (`get`, `put`, `post`, etc)
and a `Str $.name` to `RouteHandler` class.

On `Cro::HTTP::Router` it adds `Handler %.named-routes` and `Handler %.generated-named-routes`
to store the handlers indexed by name (passed or generated).

It also adds the methods `generate-name` to be able to generate a name
if one wasnt provided, and `path-template` and `path` to, having the
`Handler` (you can gat using the new exported sub `endpoint(Str $name)`)
be able to get it's path-template and it's path (receiving named
parameters to populate the vars).

It also alters `redirect` to, if the location passed is a name of an
endpoint, uses its path (it also accepts `:%params` to be used to
populate the path) as location to be redirected (unless also passed
`:!named-endpoint`).

And finally it also exports a sub called `list-routes` that returns a
string listing all routes and their names.